WiP handling of up to 5 payments
